# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TrustBridge API
  description: |
    Backend API for TrustBridge — a portable business trust layer built on IOTA.

    ## Overview
    TrustBridge lets companies create reusable trust profiles with evidence-backed signals
    and live on-chain status. Three roles interact with the system:

    - **Attester**: A country's company registry or accredited authority that verifies companies exist
    - **Company**: A business that builds a trust profile (documents, stake, deals)
    - **Verifier**: Any party that checks a company's trust profile before doing business

    ## Trust Stars (0–4)
    | Star | Name | How to earn |
    |------|------|------------|
    | ★ | Verified | Attester (authority) issues a Verifiable Credential confirming the company is registered |
    | ★★ | Staked | Company locks IOTA tokens as collateral in the escrow pool |
    | ★★★ | Proven | Company completes 3+ deals on the platform |
    | ★★★★ | Vouched | Another company vouches for this company |

    ## On-chain anchoring
    All trust events are recorded on the IOTA Rebased (Move) blockchain.
    Verifiable Credentials are issued via IOTA Identity (W3C DID + VC standards).
    An audit trail with hash-chain integrity is maintained for every profile.

  version: 0.1.0
  contact:
    name: TrustBridge Team
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.trustbridge.io
    description: Production (future)
...
paths:
...
  /api/staking/stake:
    post:
      tags: [Staking]
      summary: Stake IOTA tokens as collateral
      description: |
        Company locks IOTA tokens in the escrow pool.
        Minimum stake: 1 IOTA. Grants the ★★ Staked star.

        **This is Step 4 in the demo flow.**
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StakeRequest'
            example:
              profileId: "0xabc123..."
              amount: 10
      responses:
        '200':
          $ref: '#/components/responses/TransactionSuccess'
  /api/staking/unstake:
    post:
      tags: [Staking]
      summary: Withdraw staked tokens
      description: |
        Company withdraws their stake. Only works if not slashed.
        Removes the ★★ Staked star.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileId]
              properties:
                profileId:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/TransactionSuccess'
  /api/staking/slash:
    post:
      tags: [Staking]
      summary: Slash a company's stake
      description: |
        Admin confiscates a fraudulent company's staked tokens.
        The tokens move to the platform's slashed funds pool.
        Also slashes the trust profile (all stars → 0).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlashStakeRequest'
      responses:
        '200':
          $ref: '#/components/responses/TransactionSuccess'
...
components:
...
  responses:
...
    TransactionSuccess:
      description: Transaction executed on-chain
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              transactionDigest:
                type: string
                description: IOTA transaction digest (can be verified on explorer)
  schemas:
...
    StakeRequest:
      type: object
      required: [profileId, amount]
      properties:
        profileId:
          type: string
          description: TrustProfile object ID
        amount:
          type: number
          description: Amount in IOTA (minimum 1 IOTA)
          minimum: 1
    SlashStakeRequest:
      type: object
      required: [profileId, stakerAddress, adminCapId]
      properties:
        profileId:
          type: string
        stakerAddress:
          type: string
          description: IOTA address of the company being slashed
        adminCapId:
          type: string
...